+++ app/controllers/RobinhoodPdfController.ts
@@ -1,0 +1,7 @@
+
+
+export class RobinhoodPdfController {
+
+  
+
+}
